comment out all logic from index.js and move it to world.js
