+++ test/world-test.js
@@ -57,5 +57,7 @@ describe('World', function(){
       assert.isAbove(world.tray.x, 250);
       assert.isBelow(world.tray.x, 500);
     });
+
+    
   });
 });
